test: move non-canonicity tests to new file
