+++ plasma_framework/test/src/exits/payment/PaymentStartInFlightExit.test.js
@@ -575,239 +575,5 @@ contract('PaymentInFlightExitRouter', ([_, alice, bob, carol]) => {
                 );
             });
         });
-
-        describe('canonicity challenge', () => {
-            function createCompetitorTransaction(ifeZeroInput, otherInput) {
-                const competingTx = createInputTransaction([otherInput.utxoPos, ifeZeroInput], bob, AMOUNT);
-                const competingTxPos = new UtxoPos(buildUtxoPos(otherInput.blockNum + CHILD_BLOCK_INTERVAL, 0, 0));
-
-                return {
-                    competingTx: web3.utils.bytesToHex(competingTx.rlpEncoded()),
-                    decodedCompetingTx: competingTx,
-                    competingTxPos,
-                };
-            }
-
-            function createInclusionProof(encodedTx, txUtxoPos) {
-                const merkleTree = new MerkleTree([encodedTx], MERKLE_TREE_HEIGHT);
-                const competingTxInclusionProof = merkleTree.getInclusionProof(encodedTx);
-
-                return {
-                    competingTxInclusionProof,
-                    blockHash: merkleTree.root,
-                    blockNum: txUtxoPos.blockNum,
-                    blockTimestamp: 1000,
-                };
-            }
-
-            function createValidNoncanonicalChallengeArgs(decodedIfeTx) {
-                const utxoPos = new UtxoPos(buildUtxoPos(BLOCK_NUMBER, 2, 2));
-                const { competingTx, decodedCompetingTx, competingTxPos } = createCompetitorTransaction(
-                    decodedIfeTx.inputs[0], utxoPos,
-                );
-
-                const {
-                    competingTxInclusionProof, blockHash, blockNum, blockTimestamp,
-                } = createInclusionProof(
-                    competingTx, competingTxPos,
-                );
-
-                const competingTxWitness = addressToOutputGuard(bob);
-
-                return {
-                    args: {
-                        inFlightTx: web3.utils.bytesToHex(decodedIfeTx.rlpEncoded()),
-                        inFlightTxInputIndex: 0,
-                        competingTx,
-                        competingTxInputIndex: 1,
-                        competingTxInputOutputType: OUTPUT_TYPE_ZERO,
-                        competingTxPos: competingTxPos.utxoPos,
-                        competingTxInclusionProof,
-                        competingTxWitness,
-                    },
-                    block: {
-                        blockHash, blockNum, blockTimestamp,
-                    },
-                    decodedCompetingTx,
-                };
-            }
-
-            beforeEach(async () => {
-                this.framework = await SpyPlasmaFramework.new(
-                    MIN_EXIT_PERIOD, DUMMY_INITIAL_IMMUNE_VAULTS_NUM, INITIAL_IMMUNE_EXIT_GAME_NUM,
-                );
-                this.spendingConditionRegistry = await PaymentSpendingConditionRegistry.new();
-                this.exitGame = await PaymentInFlightExitRouter.new(
-                    this.framework.address, this.spendingConditionRegistry.address,
-                );
-
-                const {
-                    args,
-                    argsDecoded,
-                    inputTxsBlockRoot1,
-                    inputTxsBlockRoot2,
-                } = buildValidIfeStartArgs(AMOUNT, [alice, bob, carol], BLOCK_NUMBER, DEPOSIT_BLOCK_NUMBER);
-                this.args = args;
-                this.argsDecoded = argsDecoded;
-                await this.framework.setBlock(BLOCK_NUMBER, inputTxsBlockRoot1, 0);
-                await this.framework.setBlock(DEPOSIT_BLOCK_NUMBER, inputTxsBlockRoot2, 0);
-
-                const conditionTrue = await PaymentSpendingConditionTrue.new();
-
-                await this.spendingConditionRegistry.registerSpendingCondition(
-                    OUTPUT_TYPE_ZERO, IFE_TX_TYPE, conditionTrue.address,
-                );
-
-                await this.exitGame.startInFlightExit(
-                    this.args,
-                    { from: alice, value: IN_FLIGHT_EXIT_BOND },
-                );
-
-                const {
-                    args: cArgs, block, decodedCompetingTx,
-                } = createValidNoncanonicalChallengeArgs(this.argsDecoded.inFlightTx);
-
-                this.challengeArgs = cArgs;
-                this.competingTx = decodedCompetingTx;
-                this.competingTxBlock = block;
-            });
-
-            it('should successfully challenge ife', async () => {
-                await this.framework.setBlock(
-                    this.competingTxBlock.blockNum,
-                    this.competingTxBlock.blockHash,
-                    this.competingTxBlock.blockTimestamp,
-                );
-
-                const { receipt } = await this.exitGame.challengeInFlightExitNotCanonical(
-                    this.challengeArgs, { from: alice },
-                );
-
-                await expectEvent.inTransaction(
-                    receipt.transactionHash,
-                    PaymentChallengeIFENotCanonical,
-                    'InFlightExitChallenged',
-                    {
-                        challenger: alice,
-                        txHash: web3.utils.sha3(this.args.inFlightTx),
-                        challengeTxPosition: new BN(this.challengeArgs.competingTxPos),
-                    },
-                );
-            });
-
-            it('fails when competing tx is the same as in-flight one', async () => {
-                this.challengeArgs.competingTx = this.challengeArgs.inFlightTx;
-
-                await expectRevert(
-                    this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice }),
-                    'The competitor transaction is the same as transaction in-flight',
-                );
-            });
-
-            it('fails when first phase is over', async () => {
-                await time.increase((MIN_EXIT_PERIOD / 2) + 1);
-
-                await expectRevert(
-                    this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice }),
-                    'Canonicity challege phase for this exit has ended',
-                );
-            });
-
-            it('fails when competing tx is not included in the given position', async () => {
-                await expectRevert(
-                    this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice }),
-                    'Transaction is not included in block of plasma chain.',
-                );
-            });
-
-            it('fails when ife not started', async () => {
-                this.challengeArgs.inFlightTx = this.challengeArgs.competingTx;
-
-                await expectRevert(
-                    this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice }),
-                    "In-fligh exit doesn't exists",
-                );
-            });
-
-            it('fails when spending condition is not met', async () => {
-                const newOutputType = OUTPUT_TYPE_ZERO + 1;
-
-                const conditionFalse = await PaymentSpendingConditionFalse.new();
-                await this.spendingConditionRegistry.registerSpendingCondition(
-                    newOutputType, IFE_TX_TYPE, conditionFalse.address,
-                );
-
-                this.challengeArgs.competingTxInputOutputType = newOutputType;
-                await expectRevert(
-                    this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice }),
-                    'Competing input spending condition is not met',
-                );
-            });
-
-            it('fails when competing tx is younger than already known competitor', async () => {
-                // challenge ife as previously
-                await this.framework.setBlock(
-                    this.competingTxBlock.blockNum,
-                    this.competingTxBlock.blockHash,
-                    this.competingTxBlock.blockTimestamp,
-                );
-
-                await this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice });
-
-                // then mine the next block - with the same root hash
-                const nextBlockNum = this.competingTxBlock.blockNum + CHILD_BLOCK_INTERVAL;
-                const nextBlockTimestamp = this.competingTxBlock.blockTimestamp + 1000;
-                const nextCompetitorPos = buildUtxoPos(nextBlockNum, 0, 0);
-
-                await this.framework.setBlock(
-                    nextBlockNum, this.competingTxBlock.blockHash, nextBlockTimestamp,
-                );
-
-                // try to challenge again with competitor from the lastly mined block
-                this.challengeArgs.competingTxPos = nextCompetitorPos;
-                await expectRevert(
-                    this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice }),
-                    'Competing transaction is not older than already known competitor',
-                );
-            });
-
-            it('fails when challenge with the same competing tx twice', async () => {
-                await this.framework.setBlock(
-                    this.competingTxBlock.blockNum,
-                    this.competingTxBlock.blockHash,
-                    this.competingTxBlock.blockTimestamp,
-                );
-
-                await this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice });
-
-                await expectRevert(
-                    this.exitGame.challengeInFlightExitNotCanonical(this.challengeArgs, { from: alice }),
-                    'Competing transaction is not older than already known competitor',
-                );
-            });
-
-            it('should set large competitor position when competitor is in-flight', async () => {
-                this.challengeArgs.competingTxPos = 0;
-                this.challengeArgs.competingTxInclusionProof = '0x';
-
-                // it seems to be solidity `~uint256(0)` - what is important here: it's HUGE
-                const expectedCompetitorPos = new BN(2).pow(new BN(256)).sub(new BN(1));
-
-                const { receipt } = await this.exitGame.challengeInFlightExitNotCanonical(
-                    this.challengeArgs, { from: alice },
-                );
-
-                await expectEvent.inTransaction(
-                    receipt.transactionHash,
-                    PaymentChallengeIFENotCanonical,
-                    'InFlightExitChallenged',
-                    {
-                        challenger: alice,
-                        txHash: web3.utils.sha3(this.args.inFlightTx),
-                        challengeTxPosition: expectedCompetitorPos,
-                    },
-                );
-            });
-        });
     });
 });
